# Implement faster insertion