--- a/ponairi.nimble
+++ b/ponairi.nimble
@@ -9,5 +9,5 @@ srcDir        = "src"
 
 # Dependencies
 
-requires "nim >= 2.0.0"
+requires "nim >= 2.2.0"
 requires "https://github.com/ire4ever1190/lowdb#fix/conditional-dependency"