Implements faster insertion of items by reusing the prepared statement instead of recreating it for each item. From benchmarks it looks to be roughly 4x faster

Benchmark code
```nim
import benchy
import src/ponairi
import strformat
import std/random

type
  Person = object
    name: string
    age: int
    alive: bool

# I am uncreative with names
const names = ["Jake", "James", "John", "Jack"]

proc oldInsert*[T: SomeTable](db: DbConn, items: openArray[T]) =
  ## Inserts the list of items into the database.
  ## This gets ran in a transaction so if an error happens then none
  ## of the items are saved to the database
  db.transaction:
    for item in items:
      db.insert item

var people: seq[Person]
for i in 0..<10000:
  people &= Person(name: sample(names), age: rand(1..100), alive: rand(bool))

let
  db1 = newConn(":memory:")
  db2 = newConn(":memory:")

db1.create(Person)
db2.create(Person)

timeIt "Old insertion":
  db1.oldInsert(people)

timeIt "New insertion":
  db2.insert(people)
```

`nim r -d:release bench.nim`
```
   min time    avg time  std dv   runs name
  24.860 ms   26.600 ms  ±1.289   x185 Old insertion
   6.196 ms    6.713 ms  ±0.306   x737 New insertion
```

Todo

- [x]  Implement for normal inserting
- [ ] Implement for upsert